+++ share/pot/start.sh
@@ -239,6 +239,9 @@ _js_start()
 	if [ "$(_get_conf_var "$_pname" "pot.attr.procfs")" = "YES" ]; then
 		_param="$_param mount.procfs"
 	fi
+	if [ "$(_get_conf_var "$_pname" "pot.attr.fdescfs")" = "YES" ]; then
+		_param="$_param mount.fdescfs"
+	fi
 	_hostname="$( _get_conf_var $_pname host.hostname )"
 	_osrelease="$( _get_conf_var $_pname osrelease )"
 	_param="$_param name=$_pname host.hostname=$_hostname osrelease=$_osrelease"